코디 페이지 위젯 분리
